adds support for multi-release-jar | rework `Traces.java`
