+++ .github/workflows/ci-mrj.yml
@@ -1,0 +1,121 @@
+name: CI
+on:
+  pull_request:
+    branches:
+      - 3.6.x
+permissions: read-all
+jobs:
+  preliminary:
+    name: preliminary sanity checks
+    runs-on: ubuntu-20.04
+    steps:
+      - uses: actions/checkout@93ea575cb5d8a053eaa0ac8fa3b40d7e05a33cc8 # tag=v3
+        with:
+          fetch-depth: 0 #needed by spotless
+      - run: ${GITHUB_WORKSPACE}/.github/setup.sh
+        shell: bash
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'jdkfile'
+          java-version: 9.0.4
+          jdkFile: /opt/openjdk/java9/OpenJDK9U-jdk_x64_linux_hotspot_9.0.4_11.tar.gz
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'temurin'
+          java-version: |
+            21-ea
+            8
+      - uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # tag=v2
+        name: spotless (license header)
+        if: always()
+        with:
+          arguments: spotlessCheck -PspotlessFrom=origin/${{ github.base_ref }}
+      - uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # tag=v2
+        name: api compatibility
+        if: always()
+        with:
+          arguments: japicmp
+      - name: how to fix
+        if: failure()
+        # the foreground (38;5) color code 208 is orange. we also have bold, white bg (38;5;0;48;5;255m), white fg on black bg...
+        run: |
+          echo -e "\n\033[38;5;0;48;5;208m \u001b[1m How to deal with errors in preliminary job: \u001b[0m\033[0m"
+          echo "(Have a look at the steps above to see what failed exactly)"
+          echo -e "\n - \u001b[1mSpotless (license headers)\u001b[0m failures on touched java files \033[38;5;255;48;5;0m\u001b[1mcan be automatically fixed by running\u001b[0m:"
+          echo -e "   \033[38;5;0;48;5;255m ./gradlew spotlessApply \033[0m"
+          echo -e "\n - \u001b[1mAPI Compatibility\u001b[0m failures should be considered carefully and \033[38;5;255;48;5;0m\u001b[1mdiscussed with maintainers in the PR\u001b[0m"
+          echo "   If there are failures, the detail should be available in the step's log:"
+          echo -e "   Look for the \033[38;5;0;48;5;255m API compatibility failures \033[0m block(s)."
+          echo "   Alternatively, locally run the following command to get access to the full report:"
+          echo -e "   \033[38;5;0;48;5;255m ./gradlew japicmp \033[0m"
+          echo ""
+          exit -1
+  core-fast:
+    name: core fast tests
+    runs-on: ubuntu-latest
+    needs: preliminary
+    steps:
+      - uses: actions/checkout@93ea575cb5d8a053eaa0ac8fa3b40d7e05a33cc8 # tag=v3
+      - run: ${GITHUB_WORKSPACE}/.github/setup.sh
+        shell: bash
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'jdkfile'
+          java-version: 9.0.4
+          jdkFile: /opt/openjdk/java9/OpenJDK9U-jdk_x64_linux_hotspot_9.0.4_11.tar.gz
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'temurin'
+          java-version: |
+            21-ea
+            8
+      - uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # tag=v2
+        name: gradle
+        with:
+          arguments: :reactor-core:test --no-daemon -Pjunit-tags=!slow
+  core-slow:
+    name: core slower tests
+    runs-on: ubuntu-latest
+    needs: preliminary
+    steps:
+      - uses: actions/checkout@93ea575cb5d8a053eaa0ac8fa3b40d7e05a33cc8 # tag=v3
+      - run: ${GITHUB_WORKSPACE}/.github/setup.sh
+        shell: bash
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'jdkfile'
+          java-version: 9.0.4
+          jdkFile: /opt/openjdk/java9/OpenJDK9U-jdk_x64_linux_hotspot_9.0.4_11.tar.gz
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'temurin'
+          java-version: |
+            21-ea
+            8
+      - uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # tag=v2
+        name: gradle
+        with:
+          arguments: :reactor-core:test --no-daemon -Pjunit-tags=slow
+  other:
+    name: other tests
+    runs-on: ubuntu-latest
+    needs: preliminary
+    steps:
+      - uses: actions/checkout@93ea575cb5d8a053eaa0ac8fa3b40d7e05a33cc8 # tag=v3
+      - run: ${GITHUB_WORKSPACE}/.github/setup.sh
+        shell: bash
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'jdkfile'
+          java-version: 9.0.4
+          jdkFile: /opt/openjdk/java9/OpenJDK9U-jdk_x64_linux_hotspot_9.0.4_11.tar.gz
+      - uses: actions/setup-java@de1bb2b0c5634f0fc4438d7aa9944e68f9bf86cc # tag=v3
+        with:
+          distribution: 'temurin'
+          java-version: |
+            21-ea
+            8
+      - uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # tag=v2
+        name: other tests
+        with:
+          arguments: check -x :reactor-core:test -x spotlessCheck --no-daemon
