+++ build.gradle
@@ -38,12 +38,14 @@ plugins {
 	alias(libs.plugins.nohttp)
 	alias(libs.plugins.jcstress) apply false
 	alias(libs.plugins.spotless)
+	alias(libs.plugins.mrjar) apply false
 }
 
 apply plugin: "io.reactor.gradle.detect-ci"
 apply from: "gradle/asciidoc.gradle" // asciidoc (which is generated from root dir)
 apply from: "gradle/releaser.gradle"
 apply from: "gradle/dependencies.gradle"
+apply from: "gradle/toolchains.gradle"
 
 repositories { //needed at root for asciidoctor and nohttp-checkstyle
 	mavenCentral()
@@ -141,6 +143,7 @@ configure(subprojects) { p ->
 	apply plugin: 'java'
 	apply plugin: 'jacoco'
 	apply from: "${rootDir}/gradle/setup.gradle"
+    apply from: "${rootDir}/gradle/toolchains.gradle"
 
 	description = 'Non-Blocking Reactive Foundation for the JVM'
 	group = 'io.projectreactor'
