+++ reactor-core/src/main/java/reactor/core/publisher/Traces.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2021 VMware Inc. or its affiliates, All Rights Reserved.
+ * Copyright (c) 2018-2023 VMware Inc. or its affiliates, All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,14 +16,11 @@
 
 package reactor.core.publisher;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import sun.misc.JavaLangAccess;
-import sun.misc.SharedSecrets;
 
 /**
  * Utilities around manipulating stack traces and displaying assembly traces.
@@ -48,217 +45,7 @@ final class Traces {
 	 * each element being prepended with a tabulation and appended with a
 	 * newline.
 	 */
-	static Supplier<Supplier<String>> callSiteSupplierFactory;
-
-	static {
-		String[] strategyClasses = {
-				Traces.class.getName() + "$StackWalkerCallSiteSupplierFactory",
-				Traces.class.getName() + "$SharedSecretsCallSiteSupplierFactory",
-				Traces.class.getName() + "$ExceptionCallSiteSupplierFactory",
-		};
-		// find one available call-site supplier w.r.t. the jdk version to provide
-		// linkage-compatibility between jdk 8 and 9+
-		callSiteSupplierFactory = Stream
-				.of(strategyClasses)
-				.flatMap(className -> {
-					try {
-						Class<?> clazz = Class.forName(className);
-						@SuppressWarnings("unchecked")
-						Supplier<Supplier<String>> function = (Supplier) clazz.getDeclaredConstructor()
-						                                                      .newInstance();
-						return Stream.of(function);
-					}
-					// explicitly catch LinkageError to support static code analysis
-					// tools detect the attempt at finding out jdk environment
-					catch (LinkageError e) {
-						return Stream.empty();
-					}
-					catch (Throwable e) {
-						return Stream.empty();
-					}
-				})
-				.findFirst()
-				.orElseThrow(() -> new IllegalStateException("Valid strategy not found"));
-	}
-
-	/**
-	 * Utility class for the call-site extracting on Java 9+.
-	 *
-	 */
-	@SuppressWarnings("unused")
-	static final class StackWalkerCallSiteSupplierFactory implements Supplier<Supplier<String>> {
-
-		static {
-			// Trigger eager StackWalker class loading.
-			StackWalker.getInstance();
-		}
-
-		/**
-		 * Transform the current stack trace into a {@link String} representation,
-		 * each element being prepended with a tabulation and appended with a
-		 * newline.
-		 *
-		 * @return the string version of the stacktrace.
-		 */
-		@Override
-		public Supplier<String> get() {
-			StackWalker.StackFrame[] stack = StackWalker.getInstance().walk(s -> {
-				StackWalker.StackFrame[] result = new StackWalker.StackFrame[10];
-				Iterator<StackWalker.StackFrame> iterator = s.iterator();
-				iterator.next(); // .get
-
-				int i = 0;
-				while (iterator.hasNext()) {
-					StackWalker.StackFrame frame = iterator.next();
-
-					if (i >= result.length) {
-						return new StackWalker.StackFrame[0];
-					}
-
-					result[i++] = frame;
-
-					if (isUserCode(frame.getClassName())) {
-						break;
-					}
-				}
-				StackWalker.StackFrame[] copy = new StackWalker.StackFrame[i];
-				System.arraycopy(result, 0, copy, 0, i);
-				return copy;
-			});
-
-			if (stack.length == 0) {
-				return () -> "";
-			}
-
-			if (stack.length == 1) {
-				return () -> "\t" + stack[0].toString() + "\n";
-			}
-
-			return () -> {
-				StringBuilder sb = new StringBuilder();
-
-				for (int j = stack.length - 2; j > 0; j--) {
-					StackWalker.StackFrame previous = stack[j];
-
-					if (!full) {
-						if (previous.isNativeMethod()) {
-							continue;
-						}
-
-						String previousRow = previous.getClassName() + "." + previous.getMethodName();
-						if (shouldSanitize(previousRow)) {
-							continue;
-						}
-					}
-					sb.append("\t")
-					  .append(previous.toString())
-					  .append("\n");
-					break;
-				}
-
-				sb.append("\t")
-				  .append(stack[stack.length - 1].toString())
-				  .append("\n");
-
-				return sb.toString();
-			};
-		}
-	}
-
-	@SuppressWarnings("unused")
-	static class SharedSecretsCallSiteSupplierFactory implements Supplier<Supplier<String>> {
-
-		@Override
-		public Supplier<String> get() {
-			return new TracingException();
-		}
-
-		static class TracingException extends Throwable implements Supplier<String> {
-
-			static final JavaLangAccess javaLangAccess = SharedSecrets.getJavaLangAccess();
-
-			@Override
-			public String get() {
-				int stackTraceDepth = javaLangAccess.getStackTraceDepth(this);
-
-				StackTraceElement previousElement = null;
-				// Skip get()
-				for (int i = 2; i < stackTraceDepth; i++) {
-					StackTraceElement e = javaLangAccess.getStackTraceElement(this, i);
-
-					String className = e.getClassName();
-					if (isUserCode(className)) {
-						StringBuilder sb = new StringBuilder();
-
-						if (previousElement != null) {
-							sb.append("\t").append(previousElement.toString()).append("\n");
-						}
-						sb.append("\t").append(e.toString()).append("\n");
-						return sb.toString();
-					}
-					else {
-						if (!full && e.getLineNumber() <= 1) {
-							continue;
-						}
-
-						String classAndMethod = className + "." + e.getMethodName();
-						if (!full && shouldSanitize(classAndMethod)) {
-							continue;
-						}
-						previousElement = e;
-					}
-				}
-
-				return "";
-			}
-		}
-	}
-
-	@SuppressWarnings("unused")
-	static class ExceptionCallSiteSupplierFactory implements Supplier<Supplier<String>> {
-
-		@Override
-		public Supplier<String> get() {
-			return new TracingException();
-		}
-
-		static class TracingException extends Throwable implements Supplier<String> {
-
-			@Override
-			public String get() {
-				StackTraceElement previousElement = null;
-				StackTraceElement[] stackTrace = getStackTrace();
-				// Skip get()
-				for (int i = 2; i < stackTrace.length; i++) {
-					StackTraceElement e = stackTrace[i];
-
-					String className = e.getClassName();
-					if (isUserCode(className)) {
-						StringBuilder sb = new StringBuilder();
-
-						if (previousElement != null) {
-							sb.append("\t").append(previousElement.toString()).append("\n");
-						}
-						sb.append("\t").append(e.toString()).append("\n");
-						return sb.toString();
-					}
-					else {
-						if (!full && e.getLineNumber() <= 1) {
-							continue;
-						}
-
-						String classAndMethod = className + "." + e.getMethodName();
-						if (!full && shouldSanitize(classAndMethod)) {
-							continue;
-						}
-						previousElement = e;
-					}
-				}
-
-				return "";
-			}
-		}
-	}
+	static final Supplier<Supplier<String>> callSiteSupplierFactory = new CallSiteSupplierFactory();
 
 	/**
 	 * Return true for strings (usually from a stack trace element) that should be
