--- reactor-core/src/main/java8stubs/sun/misc/SharedSecrets.java
@@ -1,9 +1,0 @@
-package sun.misc;
-
-/**
- * Stub for the Java 8 compatibility when compiled with JDK 9+.
- */
-public class SharedSecrets {
-
-	public static native JavaLangAccess getJavaLangAccess();
-}
